Remove StoreAndStatus model and merge it with Store model
