+++ FosServer/src/main/java/com/example/fostracker/models/Store.java
@@ -16,6 +16,7 @@ public class Store {
     String phone;
     String storeName;
     Timestamp creationDateTime;
+    String verificationStatus;
 
     /**
      * @param ownerName   - the name of the business owner
@@ -66,6 +67,48 @@ public class Store {
         this.phone = phone;
         this.storeName = storeName;
         this.creationDateTime = creationDateTime;
+    }
+
+    /**
+     * If store verification is successful then database table stores "green" String.
+     * else if merchant is not registered then database table stores "grey" String
+     * else if store verification is  unsuccessful then database table stores "red" String
+     * else if merchant needs to be revisited then database table stores "grey" String
+     */
+    public static final String VERIFICATION_SUCCESSFUL = "green";
+    public static final String NOT_VERIFIED = "grey";
+    public static final String VERIFICATION_UNSUCCESSFUL = "red";
+    public static final String NEEDS_REVISIT = "yellow";
+    public static final int VERIFICATION_SUCCESSFUL_INT = 2;
+    public static final int NOT_VERIFIED_INT = 0;
+    public static final int VERIFICATION_UNSUCCESSFUL_INT = -1;
+    public static final int NEEDS_REVISIT_INT = 1;
+
+    /**
+     * @param storePhone              - the store phone number that was verified
+     * @param storeCoordinates        - the location the store
+     * @param verificationStatus      - whether the verification was successful/unsuccessful/incomplete
+     */
+    public Store(String storePhone, Coordinates storeCoordinates, int verificationStatus) {
+        this.phone = storePhone;
+        this.coordinates = storeCoordinates;
+        this.verificationStatus = getStoreVerificationString(verificationStatus);
+    }
+
+    //Based on the status integral value it sets to required status string
+    private String getStoreVerificationString(int verificationStatus) {
+        switch (verificationStatus) {
+            case VERIFICATION_UNSUCCESSFUL_INT:
+                return VERIFICATION_UNSUCCESSFUL;
+            case VERIFICATION_SUCCESSFUL_INT:
+                return VERIFICATION_SUCCESSFUL;
+            case NOT_VERIFIED_INT:
+                return NOT_VERIFIED;
+            case NEEDS_REVISIT_INT:
+                return NEEDS_REVISIT;
+            default:
+                return null;
+        }
     }
 
     public String getStoreName() {
@@ -119,5 +162,13 @@ public class Store {
 
     public void setOwnerName(Name ownerName) {
         this.ownerName = ownerName;
+    }
+
+    public String getStoreVerificationStatus() {
+        return verificationStatus;
+    }
+
+    public void setStoreVerificationStatus(String status) {
+        this.verificationStatus = status;
     }
 }
